added sensor interface
